src/main/java/com/onaio/steps/activityHandler/BackHomeHandler.java
package com.onaio.steps.activityHandler;

import android.app.Activity;
import android.app.ListActivity;
import android.content.Intent;

import com.onaio.steps.activity.StepsActivity;
import com.onaio.steps.activityHandler.Interface.IMenuHandler;
import com.onaio.steps.helper.Constants;

public class BackHomeHandler implements IMenuHandler {

    private Activity activity;

    public BackHomeHandler(Activity activity) {
        this.activity = activity;
    }

    @Override
    public boolean shouldOpen(int menu_id) {
        return android.R.id.home == menu_id;
    }

    @Override
    public boolean open() {
        activity.finish();
        return true;
    }
}
